+++ iam/apiv3/iampb/policy_bindings_service.pb.go
@@ -365,11 +365,9 @@ type ListPolicyBindingsRequest struct {
 	//
 	// Some examples of filter queries:
 	//
-	// | Query            | Description                                         |
-	// |------------------|-----------------------------------------------------|
-	// | `target:ex*`     | The binding target's name starts with "ex".         |
-	// | `target:example` | The binding target's name is `example`.             |
-	// | `policy:example` | The binding policy's name is `example`.             |
+	// * `target:ex*`: The binding target's name starts with "ex".
+	// * `target:example`: The binding target's name is `example`.
+	// * `policy:example`: The binding policy's name is `example`.
 	Filter string `protobuf:"bytes,4,opt,name=filter,proto3" json:"filter,omitempty"`
 }